Got all logic programming part to viewModel. In fragment I only have UI calls.
Bug / Issue -> app does not update pass strength level warning in two moments, when I turn off/on switches and when changing betwen portrait and landscape view. Though this is related to viewModel, I need to address the switches case first because I think I'll solve both problems with it.