Bug Correction: -> In landscape mode, in bigger phones, UI is not moving correctly. Stays in center not taking advantage of all available space
Notes:
Thought of passing all animations to an xml file but since I have both things, some animation in XML file and rest of it in fragment, think I'll keep it this way. The aim of this App was to learn and, this way, I have two different ways of doing it. The same applies for most of my code.